fix:#02 debugger;を挿入して確認するが、動きが違います
